pyproject.toml
[tool.poetry]
name = "wfm-broker"
version = "0.1.0"
description = "Notifications for new warframe.market orders matching criteria."
authors = ["desophos <1887450+desophos@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.7.4"
quart = "^0.18.3"
pytest-asyncio = "^0.15.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"

[tool.poetry.scripts]
start = "wfm_broker:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
